% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{addAnatomy}
\alias{addAnatomy}
\title{Add an anatomy to an AnatomyList object}
\usage{
addAnatomy(object, anatomy)
}
\arguments{
\item{object}{An object of class "AnatomyList".}

\item{anatomy}{An anatomy to be added.}
}
\value{
The modified object with the anatomy added.
}
\description{
This function adds an anatomy to an object of class "AnatomyList".
The anatomy is added to the @anatomy slot of the object.
}
