% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{addOntology}
\alias{addOntology}
\title{Add an ontology to an OntologyList object}
\usage{
addOntology(object, ontology)
}
\arguments{
\item{object}{An object of class "OntologyList".}

\item{ontology}{An ontology to be added.}
}
\value{
The modified object with the ontology added.
}
\description{
This function adds an ontology to an object of class "OntologyList".
The ontology is added to the @ontologies slot of the object.
}
