% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{Anatomy-class}
\alias{Anatomy-class}
\title{S4 class for holding anatomy information}
\description{
This class represents anatomy information and provides methods for accessing
and manipulating it.
}
\section{Slots}{

\describe{
\item{\code{id}}{The unique identifier of the anatomy.}

\item{\code{label}}{The label of the anatomy.}

\item{\code{description}}{The description of the anatomy.}
}}

