% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{AnatomyList-class}
\alias{AnatomyList-class}
\title{S4 class for holding a list of Anatomy objects}
\description{
This class represents a list of Anatomy objects, where each object
corresponds to separate anatomy information.
}
\section{Slots}{

\describe{
\item{\code{anatomy}}{A named list of Anatomy objects.}
}}

