% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_visualization.R
\name{animateMovement}
\alias{animateMovement}
\title{Animation of Fish and Stimulus Movement}
\usage{
animateMovement(
  behaviourData,
  filename = "animation.gif",
  fps = 20,
  duration = 5
)
}
\arguments{
\item{behaviourData}{A dataframe with columns: xFish, yFish, xStim,
yStim, time.}

\item{filename}{The name of the file where the animation will be saved.}
}
\value{
Saves the animation to a file.
}
\description{
This function creates an animation of the movement paths of both the fish
and the stimulus over time. It plots the x and y coordinates of the fish and
the stimulus, showing how the fish responds to the stimulus.
}
\examples{
animateMovement(behaviourData, "animation.gif")
}
