% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_annotation.R
\name{annotateIAD}
\alias{annotateIAD}
\title{Annotate IAD data}
\usage{
annotateIAD(data)
}
\arguments{
\item{data}{A dataframe of position of the fish and the stimulus over time.
The dataframe should have the following columns:
- x_fish
- y_fish
- x_stim
- y_stim
- time}
}
\value{
An annotated dataframe.
}
\description{
This function adds columns to the input dataframe containing the distance of
the fish from the stimulus, the angle of the fish relative to the stimulus,
and the time since the stimulus was presented.
}
