% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{compareGeneExpression}
\alias{compareGeneExpression}
\title{Compare Gene Expression}
\usage{
compareGeneExpression(userData, zfinData)
}
\arguments{
\item{userData}{The preprocessed user gene expression data.}

\item{zfinData}{The ZFIN reference gene expression data.}
}
\value{
A dataframe containing the comparison results.
}
\description{
This function compares user's gene expression data with ZFIN reference data.
It calculates the absolute difference in expression values for matching
genes.
}
\examples{
comparisonResults <- compareGeneExpression(userData, zfinData)
}
