% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{createAnatomyObjects}
\alias{createAnatomyObjects}
\title{Add anatomy information to a list of AnatomyList objects}
\usage{
createAnatomyObjects(anatomyInfo, anatomyLists)
}
\arguments{
\item{anatomyInfo}{A data frame of anatomy information.}

\item{anatomyLists}{A list of AnatomyList objects.}
}
\value{
A list of AnatomyList objects.
}
\description{
Add anatomy information to a list of AnatomyList objects
}
