% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{createOntologyObjects}
\alias{createOntologyObjects}
\title{Add ontology information to a list of OntologyList objects}
\usage{
createOntologyObjects(geneInfo, ontologyLists)
}
\arguments{
\item{geneInfo}{A data frame of gene information.}

\item{ontologyLists}{A list of OntologyList objects.}
}
\value{
A list of OntologyList objects.
}
\description{
Add ontology information to a list of OntologyList objects
}
