% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wikipathways.R
\name{displayPathwayById}
\alias{displayPathwayById}
\title{Display Pathway by ID}
\usage{
displayPathwayById(pathwayId)
}
\arguments{
\item{pathwayId}{A character string representing the ID of the pathway.}
}
\value{
The GPML file of the pathway.
}
\description{
This function displays a specific pathway from WikiPathways through the
Cytoscape application (https://cytoscape.org/).
}
\examples{
displayPathwayById("WP554")
}
