% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{downloadFromZfin}
\alias{downloadFromZfin}
\title{Download Data from ZFIN (Single Dataset)}
\usage{
downloadFromZfin(dataset, fileName, headers)
}
\arguments{
\item{dataset}{A string specifying the dataset to be downloaded.
Must be a valid name in the predefined dictionary of datasets.}

\item{fileName}{A string specifying the name of the file to save the
downloaded data.}

\item{headers}{A boolean indicating whether to include headers in the
download.}
}
\value{
Downloads the specified dataset and saves it as a file.
}
\description{
This function downloads data from a specified ZFIN dataset.
The dataset name is checked against a predefined list of valid datasets.
If the dataset is valid, the function proceeds to download the data.
}
\examples{
downloadFromZfin("Zebrafish Anatomy Term", "sampleData.csv", TRUE)
}
