% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{downloadFromZfinMultiple}
\alias{downloadFromZfinMultiple}
\title{Download Data from ZFIN (Multiple Datasets)}
\usage{
downloadFromZfinMultiple(data)
}
\arguments{
\item{data}{A dataframe containing the datasets to be downloaded and their
corresponding file names.}
}
\value{
Downloads the specified datasets and saves them as files.
}
\description{
This function downloads data from multiple ZFIN datasets.
}
\examples{
downloadFromZfinMultiple(
    data = data.frame(
        dataset = c("Zebrafish Anatomy Term", "Zebrafish Stage Series"),
        fileName = c("anatomy.csv", "stage.csv"),
        headers = c(TRUE, TRUE)
    )
)
}
