% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onLoad.R
\name{fetchAllPathways}
\alias{fetchAllPathways}
\title{Fetch All Pathways}
\usage{
fetchAllPathways()
}
\value{
A data frame containing all zebrafish pathways.
}
\description{
This function fetches all zebrafish pathways from WikiPathways.
}
\examples{
fetchAllPathways()
}
