% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wikipathways.R
\name{fetchPathwayById}
\alias{fetchPathwayById}
\title{Fetch Pathway by ID}
\usage{
fetchPathwayById(pathwayId, asXml = TRUE)
}
\arguments{
\item{pathwayId}{A character string representing the ID of the pathway.}

\item{asXml}{A boolean indicating whether the pathway should be returned as
XML.}
}
\value{
The GPML file of the pathway.
}
\description{
This function fetches a specific pathway from WikiPathways.
}
\examples{
fetchPathwayById("WP554")
}
