% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{filterQuality}
\alias{filterQuality}
\title{Filter out rows or columns based on a given quality measure or threshold}
\usage{
filterQuality(data, threshold)
}
\arguments{
\item{data}{A dataframe to filter.}

\item{threshold}{The threshold for quality filtering.}
}
\value{
A filtered dataframe.
}
\description{
Filter out rows or columns based on a given quality measure or threshold
}
\examples{
filterQuality(data, threshold)
}
