% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{Gene}
\alias{Gene}
\title{S4 Class Representing a Gene}
\description{
This S4 class represents a gene with various attributes
including Ensembl ID, ZFIN ID, expression levels at different developmental
stages, and ontology information.
}
\details{
S4 Class Representing a Gene

This S4 class represents a gene with various attributes including Ensembl ID,
ZFIN ID, expression levels at different developmental stages, and ontology
information.

This object will be used to keep track of the gene information and pass it
around between functions. To initialize, it only requires the Ensembl ID
passed in as a parameter. The other attributes can be added later.
}
\section{Slots}{

\describe{
\item{\code{geneSymbol}}{A character string representing the gene symbol.}

\item{\code{ensemblID}}{A character string representing the Ensembl ID of the gene.}

\item{\code{zfinID}}{A character string representing the ZFIN ID of the gene.}

\item{\code{anatomy}}{A list representing the anatomy information related to the
gene.}

\item{\code{ontology}}{A list representing the ontology information related to the
gene.}
}}

