% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample.R
\name{generateData}
\alias{generateData}
\title{Generate Sample Data for Fish and Stimulus Movement}
\usage{
generateData(time, step)
}
\arguments{
\item{time}{The total time for which to generate data.}

\item{step}{The time step between each data point.}
}
\value{
A data frame with columns: time, xFish, yFish, xStim, yStim.
}
\description{
This function generates a sample data frame representing the movement of a
fish and a stimulus over time. The stimulus moves in a circle, while the
fish moves randomly.
}
\examples{
generate_data(10, 0.1)
}
