% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{getAnatomy}
\alias{getAnatomy}
\title{Retrieve the list of anatomies from an AnatomyList object}
\usage{
getAnatomy(object)
}
\arguments{
\item{object}{An AnatomyList object}
}
\value{
A list of anatomies
}
\description{
Retrieve the list of anatomies from an AnatomyList object
}
