% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_zfin.R
\name{getGeneExpression}
\alias{getGeneExpression}
\title{Access gene expression data from ZFIN}
\usage{
getGeneExpression(geneSymbol)
}
\arguments{
\item{geneSymbol}{A character string representing the gene symbol.}
}
\value{
A list containing gene expression data.
}
\description{
Access gene expression data from ZFIN
}
\examples{
getGeneExpression("rgra")
}
