% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneInfo.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Access gene information from Ensembl Version 110}
\usage{
getGeneInfo(geneSymbols)
}
\arguments{
\item{geneSymbols}{A character vector representing gene symbols.}
}
\value{
A list of data frames of information on each gene.
}
\description{
This function retrieves gene information from Ensembl Version 110 for the
provided gene symbols. It uses the biomaRt package to access the zebrafish
gene dataset from Ensembl. The gene information is returned as a list of data
frames, where each data frame contains information on a single gene.
}
\examples{
getGeneInfo(c("raraa", "rargb"))
}
