% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{getOntologies}
\alias{getOntologies}
\title{Retrieve the list of ontologies from an OntologyList object}
\usage{
getOntologies(object)
}
\arguments{
\item{object}{An OntologyList object}
}
\value{
A list of ontologies
}
\description{
Retrieve the list of ontologies from an OntologyList object
}
