% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_zfin.R
\name{getPhenotype}
\alias{getPhenotype}
\title{Access phenotype data from ZFIN}
\usage{
getPhenotype(geneSymbol)
}
\arguments{
\item{geneSymbol}{A character string representing the gene symbol.}
}
\value{
A list containing phenotype data.
}
\description{
Access phenotype data from ZFIN
}
\examples{
getPhenotype("rargb")
}
