% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinAnatomy}
\alias{getZfinAnatomy}
\title{Access anatomy information from ZFIN}
\usage{
getZfinAnatomy(json)
}
\arguments{
\item{json}{A list containing gene ZFIN ribbon summary.}
}
\value{
A list containing anatomy information.
}
\description{
Access anatomy information from ZFIN
}
