% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinBackgroundInfo}
\alias{getZfinBackgroundInfo}
\title{Access information on the environment in which a gene is expressed. This is a
list of anatomy, stage range, and cellular component data for a gene.}
\usage{
getZfinBackgroundInfo(zfinId)
}
\arguments{
\item{zfinId}{A character string representing the gene's ZFIN ID.}
}
\value{
A list containing gene ZFIN ribbon summary.
}
\description{
Access information on the environment in which a gene is expressed. This is a
list of anatomy, stage range, and cellular component data for a gene.
}
\examples{
getZfinBackgroundInfo("ZDB-GENE-980526-284")

}
