% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinCellularComponent}
\alias{getZfinCellularComponent}
\title{Access cellular component information from ZFIN}
\usage{
getZfinCellularComponent(json)
}
\arguments{
\item{json}{A list containing gene ZFIN ribbon summary.}
}
\value{
A list containing cellular component information.
}
\description{
Access cellular component information from ZFIN
}
