% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinGeneExpression}
\alias{getZfinGeneExpression}
\title{Access gene expression data from ZFIN}
\usage{
getZfinGeneExpression(zfinId)
}
\arguments{
\item{zfinId}{A character string representing the gene's ZFIN ID.}
}
\value{
A list containing gene expression data.
}
\description{
Access gene expression data from ZFIN
}
\examples{
getZfinGeneExpression("ZDB-GENE-980526-284")
getZfinGeneExpression("ZDB-GENE-000210-1")
}
