% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinMutatationInfo}
\alias{getZfinMutatationInfo}
\title{Access mutation information from ZFIN}
\usage{
getZfinMutatationInfo(zfinId)
}
\arguments{
\item{zfinId}{A character string representing the gene's ZFIN ID.}
}
\value{
A list containing mutation information.
}
\description{
Access mutation information from ZFIN
}
\examples{
getZfinMutatationInfo("ZDB-GENE-980526-284")

}
