% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchZfin.R
\name{getZfinStageRange}
\alias{getZfinStageRange}
\title{Access stage range information from ZFIN}
\usage{
getZfinStageRange(json)
}
\arguments{
\item{json}{A list containing gene ZFIN ribbon summary.}
}
\value{
A list containing stage range information.
}
\description{
Access stage range information from ZFIN
}
