% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_visualization.R
\name{heatmapActivity}
\alias{heatmapActivity}
\title{Heatmap of Fish Activity}
\usage{
heatmapActivity(behaviourData)
}
\arguments{
\item{behaviourData}{A dataframe with columns: xFish, yFish.}
}
\value{
A heatmap of fish activity.
}
\description{
Generates a heatmap representing the frequency of the fish's positions,
indicating areas of high and low activity.
It requires the `ggplot2` library to create a 2D binning heatmap.
}
\examples{
heatmapActivity(behaviourData)
}
