% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{importAndPreprocessUserData}
\alias{importAndPreprocessUserData}
\title{Import and Preprocess User Gene Expression Data}
\usage{
importAndPreprocessUserData(filePath)
}
\arguments{
\item{filePath}{Path to the user's gene expression data file.}
}
\value{
A dataframe of preprocessed user gene expression data.
}
\description{
Import and Preprocess User Gene Expression Data
}
\examples{
userData <- importAndPreprocessUserData("path/to/user/data.csv")
}
