% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\encoding{UTF-8}
\name{importBehaviouralData}
\alias{importBehaviouralData}
\title{Import Behavioural Data}
\usage{
importBehaviouralData(filepath)
}
\arguments{
\item{filepath}{A string representing the path to the behavioural assay data
file.}
}
\value{
A `data.frame` containing behavioural assay data.
}
\description{
Imports behavioural assay data from a specified file path, expecting a
tab-delimited format. Ensures that the file exists and is readable prior
to importing the data.
}
\examples{
behaviour_data <- importBehaviouralData("inst/extdata/behavioural_data.csv")
head(behaviour_data)
}
\seealso{
\code{\link{importNeuroData}},
\code{\link{importGeneExpressionData}}
}
