% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\encoding{UTF-8}
\name{importNeuroData}
\alias{importNeuroData}
\title{Import Neuroanatomical Data}
\usage{
importNeuroData(filepath)
}
\arguments{
\item{filepath}{A string representing the path to the neuroanatomical data
file.}
}
\value{
A `data.frame` containing neuroanatomical data with appropriate
headers.
}
\description{
Reads and imports neuroanatomical data from a specified file path.
The data should be tab-delimited. This function checks for file existence
and readability before attempting to read the file.
}
\examples{
neuro_data <- importNeuroData("inst/extdata/neuroanatomical_data.tsv")
head(neuro_data)
}
\seealso{
\code{\link{importGeneExpressionData}},
\code{\link{importBehaviouralData}}
}
