% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_zfin.R
\name{isGeneSymbol}
\alias{isGeneSymbol}
\title{Verify that a gene symbol is valid}
\usage{
isGeneSymbol(geneSymbol)
}
\arguments{
\item{geneSymbol}{A character string representing the gene symbol.}
}
\value{
A boolean indicating whether the gene symbol is valid.
}
\description{
Verify that a gene symbol is valid
}
\examples{
isGeneSymbol("rara")
isGeneSymbol("rargb")
isGeneSymbol("rargb2")
}
