% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{Ontology-class}
\alias{Ontology-class}
\title{S4 class for holding ontology information}
\description{
This class represents ontology information and provides methods for accessing
and manipulating it.
}
\section{Slots}{

\describe{
\item{\code{id}}{The unique identifier of the ontology.}

\item{\code{name}}{The name of the ontology.}

\item{\code{definition}}{The definition of the ontology.}

\item{\code{namespace}}{The namespace of the ontology.}

\item{\code{description}}{The description of the ontology.}
}}

