% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\docType{class}
\name{OntologyList-class}
\alias{OntologyList-class}
\title{S4 class for holding a list of Ontology objects}
\description{
This class represents a list of Ontology objects, where each object
corresponds to a developmental stage.
}
\section{Slots}{

\describe{
\item{\code{ontologies}}{A named list of Ontology objects. The names are the
developmental stages.}
}}

