% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_visualization.R
\name{plotDistanceOverTime}
\alias{plotDistanceOverTime}
\title{Plot Distance of Fish from Stimulus Over Time}
\usage{
plotDistanceOverTime(behaviourData)
}
\arguments{
\item{behaviourData}{A dataframe with columns: xFish, yFish, xStim,
yStim, time.}
}
\value{
A plot showing the distance of the fish from the stimulus over time.
}
\description{
Calculates and plots the distance between the fish and the stimulus over
time, providing insight into the fish's response to the stimulus.
}
\examples{
plotDistanceOverTime(behaviourData)
}
