% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionVisualization.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Plot gene expression values over developmental stages}
\usage{
plotGeneExpression(data, gene)
}
\arguments{
\item{data}{A data frame with gene expression values. The data frame should
have one row per gene, and one column per developmental stage. The first
column should be the gene names.}

\item{gene}{The name of the gene to plot.}
}
\value{
A ggplot object representing the line plot.
}
\description{
Plot gene expression values over developmental stages
}
\examples{
# Load the data from a TSV but remove the first 4 lines (comments)
data <- readr::read_tsv("./inst/extdata/ERAD_query_results.tsv", skip = 4)
# Plot the expression of rara
plotGeneExpression(data, "rara")
}
