% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/behaviour_visualization.R
\name{plotTrajectory}
\alias{plotTrajectory}
\title{Plot Movement Trajectory of Fish and Stimulus}
\usage{
plotTrajectory(behaviourData)
}
\arguments{
\item{behaviourData}{A dataframe with columns: xFish, yFish, xStim,
yStim, time.}
}
\value{
A plot of the trajectory of fish and stimulus movement.
}
\description{
This function visualizes the movement paths of both the fish and the stimulus
over time.
It plots the x and y coordinates of the fish and the stimulus, showing how
the fish responds to the stimulus.
}
\examples{
plotTrajectory(behaviourData)
}
