% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{removeNA}
\alias{removeNA}
\title{Remove rows with missing values from a dataframe}
\usage{
removeNA(data)
}
\arguments{
\item{data}{A dataframe from which to remove missing values.}
}
\value{
A dataframe with missing values removed.
}
\description{
Remove rows with missing values from a dataframe
}
\examples{
removeNA(data)
}
