% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAtlaZ.R
\name{runAtlaZ}
\alias{runAtlaZ}
\title{Run the Shiny App for AtlaZ}
\usage{
runAtlaZ()
}
\description{
A function that runs the Shiny app for AtlaZ. This function is exported so
that it can be called from the command line.
}
\examples{
\dontrun{
AtlaZ::runAtlaZ()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
