% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access_wikipathways.R
\name{selectPathwayById}
\alias{selectPathwayById}
\title{Select Pathway by ID}
\usage{
selectPathwayById(pathwayId)
}
\arguments{
\item{pathwayId}{A character string representing the ID of the pathway.}
}
\value{
A data frame row containing the selected pathway.
}
\description{
This function selects a specific pathway from the list of all pathways.
}
\examples{
selectPathwayById("WP554")
}
