% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{setAnatomy}
\alias{setAnatomy}
\title{Set the anatomy for an object}
\usage{
setAnatomy(object, anatomy)

setAnatomy(object, anatomy)
}
\arguments{
\item{object}{A Gene object.}

\item{anatomy}{A list of anatomy data related to the gene.}

\item{value}{A list of anatomies to be set.}
}
\value{
The modified object with the anatomy set.

The Gene object with updated anatomy data.
}
\description{
This function sets the anatomy for an object of class "AnatomyList".
The anatomy is stored in the @anatomy slot of the object.

Updates the anatomy data of a Gene object.
}
