% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{setOntologies}
\alias{setOntologies}
\title{Set the ontologies for an object}
\usage{
setOntologies(object, value)
}
\arguments{
\item{object}{An object of class "OntologyList".}

\item{value}{A list of ontologies to be set.}
}
\value{
The modified object with the ontologies set.
}
\description{
This function sets the ontologies for an object of class "OntologyList".
The ontologies are stored in the @ontologies slot of the object.
}
