% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{setOntology}
\alias{setOntology}
\title{Set Ontology Data for a Gene Object}
\usage{
setOntology(object, ontology)
}
\arguments{
\item{object}{A Gene object.}

\item{ontology}{A list of ontology data related to the gene.}
}
\value{
The Gene object with updated ontology data.
}
\description{
Updates the ontology data of a Gene object.
}
