% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{setZfinID}
\alias{setZfinID}
\title{Set ZFIN ID for a Gene Object}
\usage{
setZfinID(object, zfinID)
}
\arguments{
\item{object}{A Gene object.}

\item{zfinID}{A character string representing the new ZFIN ID.}
}
\value{
The Gene object with updated ZFIN ID.
}
\description{
Updates the ZFIN ID of a Gene object.
}
