% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{show,Anatomy-method}
\alias{show,Anatomy-method}
\title{Show an Anatomy Object}
\usage{
\S4method{show}{Anatomy}(object)
}
\arguments{
\item{object}{An Anatomy object.}
}
\description{
Prints detailed information about an Anatomy object.
}
