% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{show,AnatomyList-method}
\alias{show,AnatomyList-method}
\title{Show an AnatomyList Object}
\usage{
\S4method{show}{AnatomyList}(object)
}
\arguments{
\item{object}{An AnatomyList object.}
}
\description{
Prints detailed information about an AnatomyList object.
}
