% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{show,Gene-method}
\alias{show,Gene-method}
\title{Show a Gene Object}
\usage{
\S4method{show}{Gene}(object)
}
\arguments{
\item{object}{A Gene object.}
}
\description{
Prints detailed information about a Gene object.
}
