% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{show,Ontology-method}
\alias{show,Ontology-method}
\title{Show an Ontology Object}
\usage{
\S4method{show}{Ontology}(object)
}
\arguments{
\item{object}{An Ontology object.}
}
\description{
Prints detailed information about an Ontology object.
}
