% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene.R
\name{show,OntologyList-method}
\alias{show,OntologyList-method}
\title{Show an OntologyList Object}
\usage{
\S4method{show}{OntologyList}(object)
}
\arguments{
\item{object}{An OntologyList object.}
}
\description{
Prints detailed information about an OntologyList object.
}
