% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{standardizeFormat}
\alias{standardizeFormat}
\title{Standardize the format of a dataframe}
\usage{
standardizeFormat(data)
}
\arguments{
\item{data}{A dataframe whose formats you wish to standardize.}
}
\value{
A dataframe with standardized formats.
}
\description{
Standardize the format of a dataframe
}
\examples{
standardizeFormat(data)
}
