% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataCleaning.R
\name{validateGeneNames}
\alias{validateGeneNames}
\title{Validate Gene Names}
\usage{
validateGeneNames(geneNames, data)
}
\arguments{
\item{geneNames}{A vector of gene names provided by the user.}

\item{data}{The dataset containing valid gene names.}
}
\value{
A character string with a message listing invalid gene names, or NULL
if all names are valid.

A character string with a message listing invalid gene names, or NULL
if all names are valid.
}
\description{
This function checks whether the provided gene names are valid based on the
loaded data.
It returns a message if invalid gene names are detected.
}
\examples{
data <- readr::read_tsv("path/to/dataset.tsv")
validateGeneNames(c("BRCA1", "INVALIDGENE"), data)
Validate Gene Names

This function checks whether the provided gene names are valid based on the
loaded data.
It returns a message if invalid gene names are detected.

data <- readr::read_tsv("path/to/dataset.tsv")
validateGeneNames(c("BRCA1", "INVALIDGENE"), data)
}
