% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchEnsembl.R
\name{getGeneInfo}
\alias{getGeneInfo}
\title{Access gene information from Ensembl Version 110}
\usage{
getGeneInfo(geneSymbol)
}
\arguments{
\item{geneSymbol}{A character string representing the gene symbol.}
}
\value{
A data frame containing gene information.
}
\description{
Access gene information from Ensembl Version 110
}
\examples{
getGeneInfo("rara")
getGeneInfo("rargb")

}
