% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\encoding{UTF-8}
\name{importGeneExpressionData}
\alias{importGeneExpressionData}
\title{Import Gene Expression Data}
\usage{
importGeneExpressionData(filepath)
}
\arguments{
\item{filepath}{A string representing the path to the gene expression data file.}
}
\value{
A `data.frame` containing gene expression data.
}
\description{
This function imports gene expression data from a file path provided.
The expected format is tab-delimited. It performs checks to ensure the
file exists and can be read before loading the data.
}
\examples{
gene_data <- importGeneExpressionData("inst/extdata/gene_expression_data.txt")
head(gene_data)
}
\seealso{
\code{\link{importNeuroData}}, \code{\link{importBehaviouralData}}
}
