% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchEnsembl.R
\name{isGeneSymbol}
\alias{isGeneSymbol}
\title{Verify that a gene symbol is valid}
\usage{
isGeneSymbol(geneSymbol, geneInfo)
}
\arguments{
\item{geneSymbol}{A character string representing the gene symbol.}

\item{geneInfo}{A data frame containing gene information.}
}
\value{
A boolean indicating whether the gene symbol is valid.
}
\description{
Verify that a gene symbol is valid
}
\examples{
geneInfo <- getGeneInfo("rara")
isGeneSymbol("rara", geneInfo)

geneInfo <- getGeneInfo("rargb")
isGeneSymbol("rargb", geneInfo)

}
