% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression_visualization.R
\name{plotGeneExpression}
\alias{plotGeneExpression}
\title{Plot gene expression values over developmental stages}
\usage{
plotGeneExpression(data, gene)
}
\arguments{
\item{data}{A data frame with gene expression values. The data frame should
have one row per gene, and one column per developmental stage. The first
column should be the gene names.}

\item{gene}{The name of the gene to plot.}
}
\value{
A ggplot object representing the line plot.
}
\description{
Plot gene expression values over developmental stages
}
\examples{
data <- read.delim("./inst/extdata/ERAD_query_results.tsv", header = TRUE,
                   stringsAsFactors = FALSE)
plotGeneExpression(data, "slc35a5")
}
